requirements.txt
fastapi
uvicorn
python-multipart
firebase-admin
spacy
scikit-learn
pandas
PyMuPDF
python-docx
nltk
pydantic
python-dotenv
